requirements.txt
Pillow
numpy
opencv-contrib-python
opencv-python
opencv-python-headless
elasticsearch
djangorestframework
django
django-cors-headers
vietocr
configparser
transformers
torch
psycopg2-binary
paddlepaddle
torchvision
paddleocr
sentence-transformers
